requirements.txt
streamlit
agno
google
google-generativeai
python-dotenv
